more fixes to the Particle and Vertex packing
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@135669 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 7a66e988e34eab462f9f4094e7a8a040257c4b0b

+++ LHCb/Event/PackedEvent/Event/PackedParticle.h
@@ -80,13 +80,13 @@ namespace LHCb
     int pmCov30,pmCov31,pmCov32;
 
     // Extra info
-    unsigned short int firstExtra, lastExtra;
+    unsigned int firstExtra, lastExtra;
 
     // End Vertex
     long long vertex;  
 
     // ProtoParticle
-    int proto;
+    long long proto;
 
     // daughters
     unsigned int firstDaughter, lastDaughter;
